Updated HDI library to version from fork

+++ lib/HDI/include/hdi/data/map_mem_eff.h
@@ -91,7 +91,7 @@ namespace hdi{
     template <class Key, class T>
     typename MapMemEff<Key,T>::mapped_type& MapMemEff<Key,T>::operator[](const key_type& k){
       int l = 0;
-      int r = _memory.size()-1;
+      int r = static_cast<int>(_memory.size())-1;
       int m = 0;
 
       if(size()==0){
@@ -128,7 +128,7 @@ namespace hdi{
     template <class Key, class T>
     typename MapMemEff<Key,T>::const_iterator MapMemEff<Key,T>::find(const key_type& k)const {
       int l = 0;
-      int r = _memory.size()-1;
+      int r = static_cast<int>(_memory.size())-1;
       int m = 0;
 
       if(size()==0){
